Divided by translation units + namespace added
